Tema07 - Proyecto Django/Clase18/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
requests